Refactor: enhance dspaceApi module to improve structure for API access
